requirements-ckiptagger.txt
absl-py==0.10.0
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1599308529326/work
backcall==0.2.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600270521970/work
blis==0.4.1
boto==2.49.0
boto3==1.14.61
botocore==1.17.61
bs4==0.0.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ckipnlp==0.9.1
ckiptagger==0.2.1
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
fake-useragent==0.1.11
future==0.18.2
gast==0.4.0
gensim==3.8.3
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1593211385768/work
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jiagu==0.2.3
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lime==0.2.0.1
lxml==4.5.2
Markdown==3.2.2
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1600368241677/work
matplotlib==3.3.2
mistune==0.8.4
mock==4.0.2
more-itertools==8.5.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1594060262917/work
networkx==2.5
nltk==3.5
notebook @ file:///Users/runner/miniforge3/conda-bld/notebook_1594379841067/work
numpy==1.19.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandas==1.1.2
pandocfilters==1.4.2
parse==1.18.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkuseg==0.0.25
plac==1.1.3
pluggy==0.13.1
praatio==4.2.0
preshed==3.0.2
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
pyee==7.0.3
Pygments==2.6.1
pylangacq==0.11.0
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1599988194096/work
pytest==6.0.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
requests-html==0.10.0
rise @ file:///Users/runner/miniforge3/conda-bld/rise_1594127187593/work
rpy2==3.3.5
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
smart-open==2.1.1
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testfixtures==6.14.2
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
treelib==1.6.1
tzlocal==2.1
urllib3==1.25.10
w3lib==1.22.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wordcloud==1.8.0
zh-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_lg-2.3.1/zh_core_web_lg-2.3.1.tar.gz
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-2.3.1/zh_core_web_sm-2.3.1.tar.gz
zipp==3.1.0
